Adds gbm surfaceless rendering and cmake build
A graphics example was needed that is platform agnostic using frameworks
that are available on most systems with a GPU. A surfaceless context was
preffered to allow flexibility on where the framebuffer is displayed.

The current example initializes EGL with a gbm surface allocated from
the GPU render node. The size can be modified through
SURFACE_WIDTH/HEIGHT preprocessor defines. Then it runs a simple
glClear, renders onto the framebuffer and exists.

Cmake was used as the build infrastructure in case more complex needs
are required with future enhancements. To build, the following should be
sufficient on a machine with all of the dependencies available:

mkdir build
cd build/
cmake ../

To cross-compile, create a similar file to the cmake-arm64.toolchain
that points to your desired filesystem. The cmake command reference:

cmake --toolchain ../cmake-arm64.toolchain ../
OR
cmake -DCMAKE_TOOLCHAIN_FILE=../cmake-arm64.toolchain ../

Signed-off-by: Erick Narvaez <e-narvaez@ti.com>